securisation acces par url

+++ index.php
@@ -1,4 +1,5 @@
 <?php
+session_start();
 //index.php fait office de routeur
 
 require("controller/controller.php"); //appel des controleurs
@@ -9,7 +10,8 @@ if (isset($_GET["action"])) // si dans l'url index une action est précisée
 {
 	if ($_GET["action"] == "post") //affichage d'un billet précis
 	{ 
-		if (isset($_GET["id"]) && $_GET["id"] > 0) {
+		if (isset($_GET["id"]) && $_GET["id"] > 0) 
+		{
 			post();
 		} else {
 			echo "Erreur : aucun id de billet transmis";
@@ -28,15 +30,27 @@ if (isset($_GET["action"])) // si dans l'url index une action est précisée
 
 	elseif ($_GET["action"] == "addPost") //ajoute un billet
 	{ 
-		newPost($_POST["auteur"], $_POST["contenu"]);
+		if(isset($_SESSION["access"]) && $_SESSION["access"] == "ok")
+		{
+			newPost($_POST["auteur"], $_POST["contenu"]);
+		} else 
+		{
+			header("Location: index.php");
+		}
 	} 
 
 	elseif ($_GET["action"] == "signal") //ajoute un signalement
 	{ 
-		if(isset($_GET["id"])){
-			newSignal();
-		} else {
-			echo "Aucun id de billet transmis";
+		if(isset($_GET["id"])&& $_GET["id"] > 0)
+		{
+			if (isset($_GET["postId"]) && $_GET["postId"] > 0)
+			{
+				newSignal();
+			}
+			else 
+			{
+				echo "Aucun id de billet transmis";
+			}
 		}
 	} 
 
@@ -50,24 +64,40 @@ if (isset($_GET["action"])) // si dans l'url index une action est précisée
 		if(isset($_POST["pseudo"]) && isset($_POST["pass"]))
 		{ 
 			access($_POST["pseudo"], $_POST["pass"]);
-		} else {
+		} else 
+		{
 			header("Location: view/connexion.php");
 		}
 	} 
 
 	elseif ($_GET["action"] == "editPost") //ouvre la page pour éditer un billet
 	{ 
-		if (isset($_GET["id"]) && $_GET["id"] > 0) {
-			editPost();
-		} else {
+		if (isset($_GET["id"]) && $_GET["id"] > 0) 
+		{
+			if(isset($_SESSION["access"]) && $_SESSION["access"] == "ok")
+			{
+				editPost();
+			} else 
+			{
+				header("Location: index.php");
+			}
+		} else 
+		{
 			echo "Aucun Id de billet transmis";
 		}
 	} 
 
 	elseif ($_GET["action"] == "changePost") //envoi le billet modifié à la bdd
 	{ 
-		if (isset($_GET["id"]) && $_GET["id"] > 0) {
-			changePost();
+		if (isset($_GET["id"]) && $_GET["id"] > 0) 
+		{
+			if(isset($_SESSION["access"]) && $_SESSION["access"] == "ok")
+			{
+				changePost();
+			} else
+			{
+				header("Location: index.php");
+			}
 		} else {
 			echo "Aucun Id de billet transmis";
 		}
@@ -75,25 +105,40 @@ if (isset($_GET["action"])) // si dans l'url index une action est précisée
 
 	elseif ($_GET["action"] == "deletePost") //supprime un billet de la bdd
 	{ 
-		if (isset($_GET["id"]) && $_GET["id"] > 0) {
-			deletePost();
-		} else {
+		if (isset($_GET["id"]) && $_GET["id"] > 0) 
+		{
+			if(isset($_SESSION["access"]) && $_SESSION["access"] == "ok")
+			{
+				deletePost();
+			} else
+			{
+				header("Location: index.php");
+			}
+		} else 
+		{
 			echo "Aucun Id de billet transmis";
 		}
 	} 
 
 	elseif ($_GET["action"] == "deleteCom") //supprime un commentaire de la bdd
 	{
-		if (isset($_GET["id"]) && $_GET["id"] > 0) {
-			deleteCom();
-		} else {
+		if (isset($_GET["id"]) && $_GET["id"] > 0) 
+		{
+			if(isset($_SESSION["access"]) && $_SESSION["access"] == "ok")
+			{
+				deleteCom();
+			} else 
+			{
+				header("Location: index.php");
+			}
+		} else 
+		{
 			echo "Aucun Id de commentaire transmis";
 		}
 	}
 
 	elseif ($_GET["action"] == "disconnect") //renvoi vers la page d'accueil et ferme session
 	{
-		session_start();
 		session_destroy();
 		header("Location: index.php");
 	}
